fix: correct type mismatches on 32-bit FreeBSD (ILP32)
Ensure WGDataIO.Size uses uint32 on ILP32 to match C struct.
Use C.size_t for nvlist_unpack and C.uint64_t for nvlist_add_number
to avoid build failures from size_t/ulong mismatches.

+++ internal/wgfreebsd/client_freebsd.go
@@ -194,7 +194,7 @@ func (c *Client) ConfigureDevice(name string, cfg wgtypes.Config) error {
 	data := wgh.WGDataIO{
 		Name: dname,
 		Data: mem,
-		Size: uint64(sz),
+		Size: uint32(sz),
 	}
 
 	if err := c.ioctlWGDataIO(wgh.SIOCSWG, &data); err != nil {
